// src/app/models/rpg/collection-item.ts
import { BtBase } from '../common/base'

export class RpgCollectionItem extends BtBase {
  pos: number
  auto: number
  name: string
  up1: string|number
  up2: string|number
  up3: string|number
  up4: string|number
  up5: string|number
  up6: string|number
  up7: string|number
  up8: string|number
  up9: string|number
  up10: string|number
  up11: string|number
  up12: string|number
  up13: string|number
  up14: string|number
  up15: string|number
  up16: string|number
  up17: string|number
  up18: string|number
  up19: string|number
  getProto() {
    return {
      pos: 0,
      auto: null,
      name: null,
      up1: null,
      up2: null,
      up3: null,
      up4: null,
      up5: null,
      up6: null,
      up7: null,
      up8: null,
      up9: null,
      up10: null,
      up11: null,
      up12: null,
      up13: null,
      up14: null,
      up15: null,
      up16: null,
      up17: null,
      up18: null,
      up19: null,
    }
  }
}